moved simd_mask_impl to a separate header.
